+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0 (Jun 19, 2026)
 * Upgraded `nullstone-io/ns` provider to `~> 0.11.0`.
 * Used `gcp_labels` from `data.ns_workspace` to label resources.
+* Used `k8s_labels` from `data.ns_workspace` for Kubernetes resource labels.
 
 # 0.1.6 (Jun 10, 2026)
 * Fixed nullstone provider upgrade.